refactor(reporting): refactored some pages & component logic and structure
