+++ package.json
@@ -28,7 +28,7 @@
 		"react": "19.0.0",
 		"react-dom": "19.0.0",
 		"react-hook-form": "^7.62.0",
-		"react-router-dom": "^7.9.1",
+		"react-router": "^7.9.2",
 		"sonner": "^2.0.7",
 		"tailwind-merge": "^2.6.0",
 		"zod": "^4.1.9"
