+++ src/components/ui/layout.tsx
@@ -1,13 +1,13 @@
 import { Toaster } from "./toast";
 
 export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <head />
-      <body>
-        <main>{children}</main>
-        <Toaster />
-      </body>
-    </html>
-  );
+	return (
+		<html lang="en">
+			<head />
+			<body>
+				<main>{children}</main>
+				<Toaster />
+			</body>
+		</html>
+	);
 }
