+++ vite.config.ts
@@ -5,19 +5,19 @@ import { defineConfig } from "vite";
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react(), tailwindcss()],
-  // server: {
-  //   proxy: {
-  //     "/api": {
-  //       target: "https://satgas.sga-cakrawala.org",
-  //       changeOrigin: true,
-  //       secure: false,
-  //     },
-  //   },
-  // },
-  resolve: {
-    alias: {
-      "@": resolve(__dirname, "./src"),
-    },
-  },
+	plugins: [react(), tailwindcss()],
+	// server: {
+	//   proxy: {
+	//     "/api": {
+	//       target: "https://satgas.sga-cakrawala.org",
+	//       changeOrigin: true,
+	//       secure: false,
+	//     },
+	//   },
+	// },
+	resolve: {
+		alias: {
+			"@": resolve(__dirname, "./src"),
+		},
+	},
 });
